+++ packages/@guardian/source-react-components/src/accordion/Accordion.tsx
@@ -21,6 +21,9 @@ export interface AccordionProps extends Props {
  * [Design System](https://theguardian.design/2a1e5182b/p/38c5aa-accordion/b/92b71e) •
  * [GitHub](https://github.com/guardian/source/blob/main/packages/@guardian/source-react-components/src/accordion/Accordion.tsx) •
  * [NPM](https://www.npmjs.com/package/@guardian/source-react-components)
+ *
+ * **Note**: the `max-height` of an `AccordionRow` is `500px`
+ *
  */
 export const Accordion = ({
 	hideToggleLabel = false,